requirements.txt
beautifulsoup4==4.10.0
Pillow==9.0.1
pandas==2.0.0
requests==2.28.2